fix(file): PR #163 코드리뷰 반영 — 락 밖에서 저장소를 확인한다
리뷰 지적 3건.

1. 락 보유 중 네트워크 I/O (FileConnectionServiceImpl)

connectAll 이 PESSIMISTIC_WRITE 락을 잡은 상태에서 파일마다 저장소를 조회했다.
Azure 면 파일당 네트워크 왕복이라, 같은 파일을 건드리는 다른 요청이 그만큼 밀린다.
확인을 먼저 하고 그다음에 락을 잡도록 순서를 뒤집었다.

이때 엔티티로 먼저 읽으면 안 된다. 영속성 컨텍스트에 올라간 뒤에는 FOR UPDATE
조회가 DB 를 다시 읽어도 1차 캐시의 인스턴스를 돌려주기 때문에, 락을 잡고 하는
재확인이 낡은 상태를 보게 된다. 필요한 값만 읽는 FileConnectionView 프로젝션을
따로 두어, 엔티티 조회가 락을 잡은 뒤의 첫 조회가 되도록 했다.

확인과 락 사이에 다른 요청이 먼저 연결할 수 있으므로 락을 잡은 뒤 다시 확인한다.
실물은 생성 전용 SAS 라 덮어쓸 수 없어 그사이 바뀌지 않는다.
이미 연결된 파일은 저장소를 확인하기 전에 걸러 헛된 왕복도 없앴다.

2. long -> int 강제 캐스팅 (TestStoredFiles)

new byte[(int) size] 는 int 범위를 넘으면 음수 길이로 조용히 실패한다.
Math.toIntExact 로 바꿔 바로 터지게 했다.

3. 포맷 (테스트 5개 파일)

일괄 치환이 남긴 콤마 뒤 공백 누락과 긴 한 줄을 여러 줄로 풀었다.
같은 치환이 어긋나게 만든 임포트 순서도 정렬했다.
spotless 는 규칙을 얕게 잡아 두어(googleJavaFormat 미적용) 이걸 잡지 못한다.

락 순서를 못으로 박는 테스트 2개를 추가했다 (InOrder 검증, 확인-락 사이 선점).

+++ src/test/java/com/getit/domain/file/TestStoredFiles.java
@@ -15,7 +15,10 @@ import org.springframework.mock.web.MockMultipartFile;
  * <p>그래서 DB 행만 만들고 연결하면 {@code FILE_NOT_UPLOADED} 로 막힌다. 연결까지 가는
  * 테스트는 이 헬퍼를 쓴다. 실제 요청 흐름에서도 업로드가 먼저이므로, 이쪽이 더 정직하다.
  *
- * <p>내용은 신고한 크기와 같은 길이의 빈 바이트다. 크기를 그대로 맞춰야
+ * <p>내용은 신고한 크기와 같은 길이의 빈 바이트다. int 범위를 넘는 크기를 넘기면
+ * {@code Math.toIntExact} 가 바로 터진다 — 테스트에서 그런 값을 쓸 일은 없다.
+ *
+ * <p>같은 길이여야 한다는 점이 중요하다. 크기를 그대로 맞춰야
  * 연결 과정의 크기 동기화가 테스트의 기대값을 바꾸지 않는다.
  */
 public final class TestStoredFiles {
@@ -33,7 +36,7 @@ public final class TestStoredFiles {
       String contentType,
       Long uploaderId) {
     fileStorage.upload(
-        new MockMultipartFile("file", originalName, contentType, new byte[(int) size]), key);
+        new MockMultipartFile("file", originalName, contentType, new byte[Math.toIntExact(size)]), key);
     return fileAssetRepository.save(
         FileAsset.upload(key, originalName, url, size, contentType, uploaderId));
   }
